+++ pkg/services/login/authinfoimpl/service.go
@@ -1,4 +1,4 @@
-package authinfoservice
+package authinfoimpl
 
 import (
 	"context"
