+++ pkg/services/login/authinfoimpl/store_test.go
@@ -1,4 +1,4 @@
-package authinfoservice
+package authinfoimpl
 
 import (
 	"context"
