+++ pkg/tests/web/index_view_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/grafana/pkg/services/login"
-	"github.com/grafana/grafana/pkg/services/login/authinfoservice"
+	"github.com/grafana/grafana/pkg/services/login/authinfoimpl"
 	"github.com/grafana/grafana/pkg/services/secrets/database"
 	secretsManager "github.com/grafana/grafana/pkg/services/secrets/manager"
 	"github.com/grafana/grafana/pkg/services/user"
@@ -129,7 +129,7 @@ func TestIntegrationIndexViewAnalytics(t *testing.T) {
 			})
 
 			secretsService := secretsManager.SetupTestService(t, database.ProvideSecretsStore(store))
-			authInfoStore := authinfoservice.ProvideStore(store, secretsService)
+			authInfoStore := authinfoimpl.ProvideStore(store, secretsService)
 
 			// insert user_auth relationship
 			err := authInfoStore.SetAuthInfo(context.Background(), &login.SetAuthInfoCommand{
